Updated names, data types and fixed(?) async methods

+++ backend/src/Models/SortHistory.cs
@@ -9,8 +9,8 @@ public class SortHistory
     public int AlgorithmId { get; set; }
     public DateTime SortStarted { get; set; }
     public DateTime SortEnded { get; set; }
-    public int TimesCompared { get; set; }
-    public int ArrayAccesses { get; set; }
+    public long TimesCompared { get; set; }
+    public long ArrayAccesses { get; set; }
     public bool WasCancelled { get; set; } = false;
 
     public TimeSpan GetTimeSpan
